update to sklearn 0.24
