+++ skdownscale/pointwise_models/__init__.py
@@ -3,6 +3,6 @@ from .bcsd import BcsdPrecipitation, BcsdTemperature
 from .core import PointWiseDownscaler
 from .gard import AnalogRegression, PureAnalog
 from .groupers import DAY_GROUPER, MONTH_GROUPER, PaddedDOYGrouper
-from .quantile_mapping import QuantileMappingReressor, TrendAwareQuantileMappingRegressor
-from .utils import LinearTrendTransformer, QuantileMapper
+from .quantile import QuantileMapper, QuantileMappingReressor, TrendAwareQuantileMappingRegressor
+from .trend import LinearTrendTransformer
 from .zscore import ZScoreRegressor
